Extract job status into a context
